legacy/evar_optimization 3/requirements.txt
absl-py @ file:///croot/absl-py_1686852429912/work
arch==7.0.0
astropy==6.0.1
astropy-iers-data==0.2024.4.15.2.45.49
box2d-py @ file:///home/conda/feedstock_root/build_artifacts/box2d-py_1695443394251/work
Brotli @ file:///tmp/abs_ecyw11_7ze/croots/recipe/brotli-split_1659616059936/work
certifi @ file:///croot/certifi_1707229174982/work/certifi
charset-normalizer @ file:///tmp/build/80754af9/charset-normalizer_1630003229654/work
clarabel==0.7.1
cloudpickle @ file:///croot/cloudpickle_1683040006038/work
contourpy @ file:///croot/contourpy_1700583582875/work
cvxpy==1.4.3
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
ecos==2.0.13
fonttools @ file:///croot/fonttools_1713551344105/work
glfw @ file:///home/conda/feedstock_root/build_artifacts/pyglfw_1708776861344/work
gym @ file:///home/conda/feedstock_root/build_artifacts/gym_1673317490724/work
gym-notices @ file:///home/conda/feedstock_root/build_artifacts/gym-notices_1671541680412/work
idna @ file:///croot/idna_1666125576474/work
imageio @ file:///croot/imageio_1707247282708/work
imageio-ffmpeg @ file:///home/conda/feedstock_root/build_artifacts/imageio-ffmpeg_1694632425602/work
importlib-metadata @ file:///croot/importlib_metadata-suite_1704813515092/work
importlib-resources @ file:///croot/importlib_resources-suite_1704281845041/work
joblib==1.4.0
kiwisolver @ file:///croot/kiwisolver_1672387140495/work
lz4 @ file:///croot/lz4_1686057876374/work
matplotlib @ file:///croot/matplotlib-suite_1713336378214/work
moviepy @ file:///home/conda/feedstock_root/build_artifacts/moviepy_1665160419595/work
mujoco @ file:///home/conda/feedstock_root/build_artifacts/mujoco_1655878372191/work/python/dist/mujoco-2.2.0.tar.gz
networkx==3.2.1
numpy @ file:///croot/numpy_and_numpy_base_1708638617955/work/dist/numpy-1.26.4-cp39-cp39-linux_x86_64.whl#sha256=b69ac3eb7538c5a224df429dc49031914fb977825ee007f2c77a13f7aa6cd769
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.7.0
osqp==0.6.5
packaging @ file:///croot/packaging_1710807400464/work
pandas==2.2.2
patsy==0.5.6
pillow @ file:///croot/pillow_1707233021655/work
ply==3.11
proglog==0.1.9
pybind11==2.12.0
pyerfa==2.0.1.4
pygame @ file:///home/conda/feedstock_root/build_artifacts/pygame_1676413338463/work
PyOpenGL==3.1.1a1
pyparsing @ file:///opt/conda/conda-bld/pyparsing_1661452539315/work
PyQt5==5.15.10
PyQt5-sip @ file:///croot/pyqt-split_1698769088074/work/pyqt_sip
PySocks @ file:///tmp/build/80754af9/pysocks_1605305812635/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytz==2024.1
PyYAML==6.0.1
qdldl==0.1.7.post2
requests @ file:///croot/requests_1707355572290/work
Riskfolio-Lib==6.0.0
scikit-learn==1.4.2
scipy==1.11.4
scs==3.2.4.post1
sip @ file:///croot/sip_1698675935381/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
statsmodels==0.14.2
threadpoolctl==3.4.0
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==1.13.1
tornado @ file:///croot/tornado_1696936946304/work
tqdm @ file:///croot/tqdm_1679561862951/work
typing_extensions==4.11.0
tzdata==2024.1
unicodedata2 @ file:///croot/unicodedata2_1713212950228/work
urllib3 @ file:///croot/urllib3_1707770551213/work
XlsxWriter==3.2.0
zipp @ file:///croot/zipp_1704206909481/work
